Add a custom error handler (#4)
This allows for a custom error handler that can be used to translate
application-specific errors into restful errors. For example, you can
use this to translate a custom "not found" error into a simple HTTP 404
Not Found. This is useful in that you can have your application return
normal application errors without having to worry about trying to
translate them into REST API errors, and just have a simple translation
layer that does all the translating.